+++ js/src/forum/components/EventFragment.js
